app/requirements.txt
flask
pyOpenSSL
idna
